+++ openmdao/core/tests/test_auto_ivc.py
@@ -248,140 +248,3 @@ class SerialTests(unittest.TestCase):
 class MPITests(SerialTests):
 
     N_PROCS = 2
-
-
-class C1(om.ExplicitComponent):
-    def setup(self):
-        self.add_input('diameter', 0.0, units='m', src_indices=[0])
-        self.add_output('z_start', 0.0, units='m')
-
-    def compute(self, inputs, outputs):
-        pass
-
-
-class C2(om.ExplicitComponent):
-
-    def setup(self):
-        self.add_input('diameter', np.zeros(3), units='m')
-
-    def compute(self, inputs, outputs):
-        pass
-
-
-class SrcIndicesTests(unittest.TestCase):
-
-    def test_mixed_src_indices_no_src_indices(self):
-        # this test passes if setup doesn't raise an exception.
-        # C1 has src_indices and C2 doesn't.
-        prob = om.Problem()
-        prob.model.add_subsystem('C1', C1(), promotes=['diameter'])
-        prob.model.add_subsystem('C2', C2(), promotes=['diameter'])
-
-        prob.setup()
-
-        prob.run_model()
-
-    def test_flat_src_inds_2_levels(self):
-        # this test passes if it doesn't raise an exception.
-        class Burn1(om.Group):
-            def setup(self):
-                self.add_subsystem('comp1', om.ExecComp(['y1=x*2'], y1=np.ones(4), x=np.ones(4)),
-                                promotes_outputs=['*'])
-
-                self.add_subsystem('comp2', om.ExecComp(['y2=x*2'], y2=np.ones(4), x=np.ones(4)),
-                                promotes_outputs=['*'])
-
-            def configure(self):
-                self.promotes('comp1', inputs=[('x', 'design:x')],
-                            src_indices=[0, 0, 0, 0], flat_src_indices=True)
-
-                self.set_input_defaults('design:x', 75.3)
-
-
-        class Traj(om.Group):
-            def setup(self):
-                self.add_subsystem('burn1', Burn1(),
-                                promotes_outputs=['*'])
-
-            def configure(self):
-                self.promotes('burn1', inputs=['design:x'],
-                            src_indices=[0, 0, 0, 0], flat_src_indices=True)
-
-        prob = om.Problem(model=Traj())
-
-        prob.setup()
-        prob.run_model()
-
-    def test_src_inds_2_subs(self):
-        # this test passes if it doesn't raise an exception.
-
-        class RHS(om.Group):
-
-            def initialize(self):
-                self.options.declare('size', 1)
-
-            def setup(self):
-                size = self.options['size']
-                self.add_subsystem('comp1', om.ExecComp(['y1=x*2'], y1=np.ones(size), x=np.ones(size)),
-                                promotes_inputs=['*'], promotes_outputs=['*'])
-
-                # test with second absolute path for 'x'
-                self.add_subsystem('comp2', om.ExecComp(['y2=x*2'], y2=np.ones(size), x=np.ones(size)),
-                                promotes_inputs=['*'], promotes_outputs=['*'])
-
-
-        class Phase(om.Group):
-            def setup(self):
-                self.add_subsystem('rhs', RHS(size=4))
-
-            def configure(self):
-                self.promotes('rhs', inputs=[('x', 'design:x')],
-                            src_indices=[0, 0, 0, 0], flat_src_indices=True)
-
-                self.set_input_defaults('design:x', 75.3)
-
-
-        class Traj(om.Group):
-            def setup(self):
-                self.add_subsystem('src', om.ExecComp(['q = b*3']))
-
-                self.add_subsystem('phase', Phase())
-
-                self.connect('src.q', 'phase.design:x')
-
-
-        prob = om.Problem(model=Traj())
-
-        prob.setup()
-
-        prob.run_model()
-
-    def test_sub_sub_promotes(self):
-        # this test passes if it doesn't raise an exception.
-        class Burn1(om.Group):
-            def setup(self):
-                self.add_subsystem('comp1', om.ExecComp(['y1=x*2'], y1=np.ones(4), x=np.ones(4)),
-                                promotes_outputs=['*'])
-                self.add_subsystem('comp2', om.ExecComp(['y2=x*2'], y2=np.ones(4), x=np.ones(4)),
-                                promotes_outputs=['*'])
-
-            def configure(self):
-                self.promotes('comp1', inputs=[('x', 'design:x')], src_indices=[0, 0, 0, 0], flat_src_indices=True)
-                self.set_input_defaults('design:x', 75.3)
-
-        class Phases(om.ParallelGroup):
-            def setup(self):
-                self.add_subsystem('burn1', Burn1(), promotes_outputs=['*'])
-
-        class Traj(om.Group):
-            def setup(self):
-                self.add_subsystem('phases', Phases(), promotes_outputs=['*'])
-
-            def configure(self):
-                # this promotes was leaving a leftover entry in _group_inputs that resulted
-                # from an earlier _setup_var_data call where the input in question had a different
-                # promoted name.
-                self.phases.promotes('burn1', inputs=['design:x'])
-
-        prob = om.Problem(model=Traj())
-        prob.setup()
